Move function definitions to cpp file
